chore: Move the License domain out TenantConfiguration
