+++ app/server/appsmith-server/src/main/java/com/appsmith/server/constants/LicensePlan.java
@@ -1,0 +1,5 @@
+package com.appsmith.server.constants;
+
+public enum LicensePlan {
+    FREE
+}
