--- app/server/appsmith-server/src/main/java/com/appsmith/server/constants/ce/LicensePlan.java
@@ -1,5 +1,0 @@
-package com.appsmith.server.constants.ce;
-
-public enum LicensePlan {
-   FREE
-}
