+++ app/server/appsmith-server/src/main/java/com/appsmith/server/domains/License.java
@@ -1,0 +1,9 @@
+package com.appsmith.server.domains;
+
+import com.appsmith.server.domains.ce.LicenseCE;
+import lombok.Data;
+
+@Data
+public class License extends LicenseCE {
+
+}
