+++ app/server/appsmith-server/src/main/java/com/appsmith/server/domains/ce/LicenseCE.java
@@ -1,0 +1,9 @@
+package com.appsmith.server.domains.ce;
+
+import com.appsmith.server.constants.LicensePlan;
+import lombok.Data;
+
+@Data
+public class LicenseCE {
+    LicensePlan plan;
+}
